+++ src/Client/ClientCore/Rendering/Scenes/Game/Gui/ResourceEditor/TileSpriteEditorTab.cs
@@ -492,6 +492,11 @@ public class TileSpriteEditorTab
             for (var i = 0; i < 3; ++i) ImGui.Spacing();
             clicked = ImGui.Button($"Any##{rowIndex}{orientation}");
         }
+        else if (neighborId == TileSprite.Empty)
+        {
+            for (var i = 0; i < 3; ++i) ImGui.Spacing();
+            clicked = ImGui.Button($"Empty##{rowIndex}{orientation}");
+        }
         else
         {
             clicked = guiExtensions.TileSpriteButton($"ctx{rowIndex}{orientation}",